% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adjust.date}
\alias{adjust.date}
\alias{adjust.next}
\alias{adjust.previous}
\title{Adjusts the given dates to the next/previous business day}
\usage{
adjust.next(dates, cal = bizdays.options$get("default.calendar"))

adjust.previous(dates, cal = bizdays.options$get("default.calendar"))
}
\arguments{
\item{dates}{dates to be adjusted}

\item{cal}{an instance of \code{Calendar}}
}
\description{
If the given dates are business days it returns the given dates, but once it
is not, it returns the next/previous business days.
}
\examples{
data(holidaysANBIMA)
cal <- Calendar(holidaysANBIMA, weekdays=c("saturday", "sunday"))
adjust.next("2013-01-01", cal)
adjust.previous("2013-01-01", cal)
}

