% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bizdayse}
\alias{bizdayse}
\alias{bizdayse.Date}
\alias{bizdayse.default}
\alias{bizyearse}
\alias{bizyearse.Date}
\alias{bizyearse.default}
\title{Business days and current days equivalence}
\usage{
bizdayse(dates, curd, cal)

\method{bizdayse}{default}(dates, curd,
  cal = bizdays.options$get("default.calendar"))

\method{bizdayse}{Date}(dates, curd,
  cal = bizdays.options$get("default.calendar"))

bizyearse(dates, curd, cal)

\method{bizyearse}{default}(dates, curd,
  cal = bizdays.options$get("default.calendar"))

\method{bizyearse}{Date}(dates, curd,
  cal = bizdays.options$get("default.calendar"))
}
\arguments{
\item{dates}{the initial date (or a vector of dates)}

\item{curd}{the amount of current days (or a vector of numeric)}

\item{cal}{an instance of Calendar}
}
\description{
\code{bizdayse} stands for business days equivalent, it returns the amount
of business days equivalent to a given number of current days.
}
\examples{
data(holidaysANBIMA)
cal <- Calendar(holidaysANBIMA, weekdays=c('saturday', 'sunday'), dib=252)
bizdayse("2013-01-02", 3, cal)
bizyearse("2013-01-02", 3, cal)
}

