% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar-holidays-weekdays}
\alias{calendar-holidays-weekdays}
\alias{holidays}
\alias{holidays.default}
\alias{holidays.Calendar}
\alias{holidays.character}
\alias{weekdays.default}
\alias{weekdays.Calendar}
\alias{weekdays.character}
\title{Calendar's holidays and weekdays}
\usage{
holidays(cal)

\method{holidays}{default}(cal)

\method{holidays}{Calendar}(cal)

\method{holidays}{character}(cal)

\method{weekdays}{default}(x, ...)

\method{weekdays}{Calendar}(x, ...)

\method{weekdays}{character}(x, ...)
}
\arguments{
\item{cal}{character with calendar name or the calendar object}

\item{x}{character with calendar name or the calendar object}

\item{...}{unused argument (this exists to keep compliance with
\code{weekdays} generic)}
}
\description{
Returns calendar's list of holidays and weekdays
}
\examples{
holidays("actual")
weekdays("actual")
# empty calls return the default calendar attributes
holidays()
weekdays()
}
