% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{holidaysB3}
\alias{holidaysB3}
\title{B3's holidays list}
\format{
a vector with \code{Date} objects that represent holidays
}
\description{
A dataset containing the list of holidays delivered by B3
(The Brazilian Exchange - https://b3.com.br)
}
\keyword{datasets}
