% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-buildin-calendars.R
\name{load_builtin_calendars}
\alias{load_builtin_calendars}
\title{Load builtin calendars}
\usage{
load_builtin_calendars()
}
\value{
Has no return
}
\description{
bizdays comes with builtins calendars:
}
\details{
- actual
- weekends
- Brazil/ANBIMA
- Brazil/B3

This function creates and registers these calendars.
Once the calendars are loaded they can be used directly by their names.

This function is called in package `.onAttach`, so it is not necessary to
call it directly.
It is for internal use, package development or in situations where the
user wants to call bizdays functions without attach the package.
}
\examples{
bizdays::load_builtin_calendars()
bizdays::calendars()
bizdays::is.bizday("2020-01-01", "Brazil/ANBIMA")
}
