% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setdiff.Date}
\alias{intersect.Date}
\alias{setdiff.Date}
\alias{setequal.Date}
\alias{union.Date}
\title{Set Operations
replace as.vector invoke with as.Date}
\usage{
setdiff.Date(x, y)

setequal.Date(x, y)

union.Date(x, y)

intersect.Date(x, y)
}
\description{
Set Operations
replace as.vector invoke with as.Date
}

