% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizdays.options.R
\docType{data}
\name{bizdays.options}
\alias{bizdays.options}
\title{bizdays' options}
\format{A \code{list} object with \emph{methods} \code{get} and \code{set} attached to.}
\usage{
bizdays.options
}
\description{
\code{bizdays.options} defines option parameters used internally in \code{bizdays}.
}
\details{
Parameters are stored in \code{bizdays.options} using \code{get} and \code{set}

\preformatted{
bizdays.options$set(option.key=value)
bizdays.options$get("option.key")
}

\code{bizdays} supports the following parameter: 

\itemize{
\item{\code{default.calendar}: }{the default calendar to be used with the
functions: \code{bizdays}, \code{bizdayse}, \code{adjust.next},
\code{adjust.previous}, \code{is.bizday}, \code{bizseq}, \code{offset}.}
}
}
\examples{
create.calendar(name="actual")
bizdays.options$set(default.calendar="actual")
bizdays("2013-07-12", "2013-07-22")

}
\keyword{datasets}

