% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bizdays}
\alias{bizdays}
\alias{bizdays.Date}
\alias{bizdays.default}
\title{Computes business days between two dates.}
\usage{
bizdays(from, to, cal)

\method{bizdays}{default}(from, to,
  cal = bizdays.options$get("default.calendar"))

\method{bizdays}{Date}(from, to,
  cal = bizdays.options$get("default.calendar"))
}
\arguments{
\item{from}{the initial date (or a vector of dates)}

\item{to}{the final date (or a vector of dates).}

\item{cal}{an instance of Calendar}
}
\description{
This function computes the amount of business days between 2 taking into
account the holidays passed to the Calendar function.
}
\examples{
data(holidaysANBIMA)
cal <- Calendar(holidaysANBIMA, weekdays=c("saturday", "sunday"))
bizdays("2013-01-02", "2013-01-31", cal)
# Once you have a default calendar set
bizdays.options$set(default.calendar=cal)
bizdays("2013-01-02", "2013-01-31")
}

