% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{holidaysANBIMA}
\alias{holidaysANBIMA}
\title{ANBIMA's holidays list}
\format{a vector with \code{Date} objects that represent holidays}
\description{
A dataset containing the list of holidays delivered by ANBIMA
(www.anbima.com.br).
}
\keyword{datasets}

