% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{holidaysANBIMA}
\alias{holidaysANBIMA}
\title{ANBIMA's holidays list}
\format{
a vector with \code{Date} objects that represent holidays
}
\description{
A dataset containing the list of holidays delivered by ANBIMA
(www.anbima.com.br).
}
\keyword{datasets}
