% Generated by roxygen2 (4.0.2): do not edit by hand
\name{offset}
\alias{add.bizdays}
\alias{offset}
\title{Offsets the given \code{dates} by \code{n} business days}
\usage{
offset(dates, n, cal = bizdays.options$get("default.calendar"))

add.bizdays(dates, n, cal = bizdays.options$get("default.calendar"))
}
\arguments{
\item{dates}{dates to be offset}

\item{n}{the amount of business days to offset}

\item{cal}{an instance of \code{Calendar}}
}
\description{
Returns the given \code{dates} offset by the given amount of \code{n}
business days.
}
\details{
The argument \code{n} accepts a sequence of integers and if its length
differs from \code{dates}' length, the recycle rule is applied to fulfill the
gap.
}
\examples{
data(holidaysANBIMA)
cal <- Calendar(holidaysANBIMA, weekdays=c("saturday", "sunday"))
offset("2013-01-02", 5, cal)
dates <- seq(as.Date("2013-01-01"), as.Date("2013-01-05"), by="day")
offset(dates, 1, cal)
}

